Optimisation du code pour la future IA + base de l'IA
- le type de donnée de la partie logique de la matrice est passé de int
(4 octets) à uint8_t (1 octet)

- la partie logique et la partie graphique de la matrice de jeu a été
séparé en 2 classes :
> LogicalTetrisBoard : partie logique
> TetrisBoard : partie graphique, héritant de sf::Drawable et de
LogicalTetrisBoard
la classe Game se sers toujours de TetrisBoard tandis que l'IA se
servira de LogicalTetrisBoard.

- On peut activer l'IA en appuyant sur F8. de F9 à F12, on règle le
nombre de tetromino sur lequel l'IA se base (attention, temps de calcul
exponentielle en fonction du nombre de tetromino connu). Infos dans
l'écran de débug

- On ne peut pas jouer en même temps que l'IA

---> Attention : l'IA ne fonctionne pas encore <---
Il s'agit ici de préparer le jeu à l'implémentation de l'IA

+++ Cpp/OpenGL_Manager.cpp
@@ -4,7 +4,7 @@ OpenGL_Manager::OpenGL_Manager(GameConfiguration* gameConfig){
     //_tetrisBoard = NULL;
 
     _gameConfig = gameConfig;
-    _tetrisBoard = new int*[BOARD_WIDTH];
+    _tetrisBoard = new uint8_t*[BOARD_WIDTH];
     orientation_vitesse = orientation_timeMax = Vector3f(0,0,0);
     orientation_progress = _gameConfig->get3DInclinaison();
     for(int i=0;i<4;i++)
@@ -12,7 +12,7 @@ OpenGL_Manager::OpenGL_Manager(GameConfiguration* gameConfig){
 
 }
 
-void OpenGL_Manager::setTetrisBoard(int tetrisBoard[][BOARD_HEIGHT]){
+void OpenGL_Manager::setTetrisBoard(uint8_t tetrisBoard[][BOARD_HEIGHT]){
     for(int i=0;i<BOARD_WIDTH;++i)
             _tetrisBoard[i]=tetrisBoard[i];
